[BUGFI] master: corrected sonar logging

+++ Acopter32-STM32F4/ArduCopter/sensors.pde
@@ -42,7 +42,7 @@ static int16_t read_sonar(void)
         return 0;
     }
 
-    int16_t temp_alt = sonar->read();
+    int32_t temp_alt = sonar->read();
 
     if (temp_alt >= sonar->min_distance && temp_alt <= sonar->max_distance * 0.70f) {
         if ( sonar_alt_health < SONAR_ALT_HEALTH_MAX ) {
